+++ webrender/src/picture.rs
@@ -28,8 +28,9 @@ use crate::prim_store::{get_raster_rects, PrimitiveScratchBuffer, RectangleKey};
 use crate::prim_store::{OpacityBindingStorage, ImageInstanceStorage, OpacityBindingIndex};
 use crate::print_tree::PrintTreePrinter;
 use crate::render_backend::DataStores;
+use crate::render_task_graph::RenderTaskId;
 use crate::render_task::{ClearMode, RenderTargetKind, RenderTask};
-use crate::render_task::{RenderTaskId, RenderTaskLocation, BlurTaskCache};
+use crate::render_task::{RenderTaskLocation, BlurTaskCache};
 use crate::resource_cache::ResourceCache;
 use crate::scene::SceneProperties;
 use crate::scene_builder::Interners;
