+++ webrender/src/renderer.rs
@@ -77,16 +77,17 @@ use crate::device::query::{GpuProfiler, GpuDebugMethod};
 use rayon::{ThreadPool, ThreadPoolBuilder};
 use crate::record::ApiRecordingReceiver;
 use crate::render_backend::{FrameId, RenderBackend};
-use crate::render_task::{RenderTargetKind, RenderTask, RenderTaskData, RenderTaskKind, RenderTaskGraph};
+use crate::render_task_graph::RenderTaskGraph;
+use crate::render_task::{RenderTargetKind, RenderTask, RenderTaskData, RenderTaskKind};
 use crate::resource_cache::ResourceCache;
 use crate::scene_builder::{SceneBuilder, LowPrioritySceneBuilder};
 use crate::screen_capture::AsyncScreenshotGrabber;
 use crate::shade::{Shaders, WrShaders};
 use smallvec::SmallVec;
 use crate::texture_cache::TextureCache;
-use crate::tiling::{AlphaRenderTarget, ColorRenderTarget, PictureCacheTarget};
-use crate::tiling::{BlitJob, BlitJobSource, RenderPassKind, RenderTargetList};
-use crate::tiling::{Frame, RenderTarget, TextureCacheRenderTarget};
+use crate::render_task_graph::{AlphaRenderTarget, ColorRenderTarget, PictureCacheTarget};
+use crate::render_task_graph::{BlitJob, BlitJobSource, RenderPassKind, RenderTargetList};
+use crate::render_task_graph::{Frame, RenderTarget, TextureCacheRenderTarget};
 use crate::util::drain_filter;
 
 use std;
